SR-11: Edit and Delete Tax Registration / Rates
